refactor(auth): separate type definition
